fixing SPI-761
git-svn-id: svn+ssh://svn.cern.ch/reps/lcgsoft/trunk/lcgcmake@7035 4525493e-7705-40b1-a816-d608a930855b

+++ pyexternals/Python_postinstall.sh
@@ -1,12 +1,14 @@
-#!/bin/sh
+#!/bin/bash
 #post-install macro to make Python-executables relocatable 
 
 # replace absolute pythonpath by relative one
 location="\#\!\/usr\/bin\/env python"
 
 for line in `find $1/bin -iname '*'`; do
-  echo "Patching $line for relocatability"
-  sed -i.bak -e "1s/.*/$location/" $line
-  rm $line.bak 
+  if head -n1 $line | grep -q "/python" ; then
+    echo "Patching $line for relocatability"
+    sed -i.bak -e "1s/.*/$location/" $line
+    rm $line.bak 
+  fi
 done
